refactor: add topics to summary numbers
